refactoring/code cleanup
