+++ src/Saml.php
@@ -14,23 +14,27 @@ use craft\events\RegisterUrlRulesEvent;
 use craft\helpers\UrlHelper;
 use craft\services\Fields;
 use craft\web\UrlManager;
+use flipbox\saml\core\AbstractPlugin;
 use flipbox\saml\core\models\SettingsInterface;
 use flipbox\saml\core\SamlPluginInterface;
-use flipbox\saml\core\AbstractPlugin;
+use flipbox\saml\core\services\Session;
 use flipbox\saml\sp\fields\ExternalIdentity;
 use flipbox\saml\sp\models\Settings;
-use flipbox\saml\sp\services\Cp;
+use flipbox\saml\sp\records\ProviderIdentityRecord;
+use flipbox\saml\sp\records\ProviderRecord;
+use flipbox\saml\sp\services\bindings\Factory;
+use flipbox\saml\sp\services\bindings\HttpPost;
+use flipbox\saml\sp\services\bindings\HttpRedirect;
+use flipbox\saml\sp\services\Login;
+use flipbox\saml\sp\services\login\User;
+use flipbox\saml\sp\services\login\UserGroups;
 use flipbox\saml\sp\services\messages\AuthnRequest;
 use flipbox\saml\sp\services\messages\LogoutRequest;
 use flipbox\saml\sp\services\messages\LogoutResponse;
 use flipbox\saml\sp\services\messages\Metadata;
 use flipbox\saml\sp\services\messages\Response;
-use flipbox\saml\sp\services\bindings\HttpPost;
-use flipbox\saml\sp\services\bindings\HttpRedirect;
-use flipbox\saml\sp\services\Login;
 use flipbox\saml\sp\services\Provider;
 use flipbox\saml\sp\services\ProviderIdentity;
-use flipbox\saml\core\services\Session;
 use yii\base\Event;
 
 /**
@@ -102,7 +106,10 @@ class Saml extends AbstractPlugin implements SamlPluginInterface
                 'authnRequest'     => AuthnRequest::class,
                 'httpPost'         => HttpPost::class,
                 'httpRedirect'     => HttpRedirect::class,
+                'bindingFactory'   => Factory::class,
                 'login'            => Login::class,
+                'user'             => User::class,
+                'userGroups'       => UserGroups::class,
                 'logoutRequest'    => LogoutRequest::class,
                 'logoutResponse'   => LogoutResponse::class,
                 'provider'         => Provider::class,
@@ -110,7 +117,6 @@ class Saml extends AbstractPlugin implements SamlPluginInterface
                 'metadata'         => Metadata::class,
                 'response'         => Response::class,
                 'session'          => Session::class,
-                'cp'               => Cp::class,
             ]
         );
     }
@@ -252,6 +258,28 @@ class Saml extends AbstractPlugin implements SamlPluginInterface
 
     /**
      * @noinspection PhpDocMissingThrowsInspection
+     * @return User
+     */
+    public function getUser()
+    {
+        /** @noinspection PhpUnhandledExceptionInspection */
+        /** @noinspection PhpIncompatibleReturnTypeInspection */
+        return $this->get('user');
+    }
+
+    /**
+     * @noinspection PhpDocMissingThrowsInspection
+     * @return UserGroups
+     */
+    public function getUserGroups()
+    {
+        /** @noinspection PhpUnhandledExceptionInspection */
+        /** @noinspection PhpIncompatibleReturnTypeInspection */
+        return $this->get('userGroups');
+    }
+
+    /**
+     * @noinspection PhpDocMissingThrowsInspection
      * @return Session
      * @throws \yii\base\InvalidConfigException
      */
@@ -263,11 +291,38 @@ class Saml extends AbstractPlugin implements SamlPluginInterface
     }
 
     /**
+     * @return Factory
+     * @throws \yii\base\InvalidConfigException
+     */
+    public function getBindingFactory()
+    {
+        /** @noinspection PhpUnhandledExceptionInspection */
+        /** @noinspection PhpIncompatibleReturnTypeInspection */
+        return $this->get('bindingFactory');
+    }
+
+    /**
      * Util Methods
      */
 
     public function getMyType()
     {
         return static::SP;
+    }
+
+    /**
+     * @return string
+     */
+    public function getProviderRecordClass()
+    {
+        return ProviderRecord::class;
+    }
+
+    /**
+     * @return string
+     */
+    public function getProviderIdentityRecordClass()
+    {
+        return ProviderIdentityRecord::class;
     }
 }
