+++ src/controllers/LogoutController.php
@@ -11,9 +11,9 @@ namespace flipbox\saml\sp\controllers;
 use craft\web\Request;
 use flipbox\saml\core\controllers\messages\AbstractLogoutController;
 use flipbox\saml\core\records\ProviderInterface;
-use flipbox\saml\core\SamlPluginInterface;
 use flipbox\saml\sp\Saml;
 use flipbox\saml\sp\services\bindings\Factory;
+use flipbox\saml\sp\traits\SamlPluginEnsured;
 use LightSaml\Model\Protocol\SamlMessage;
 use LightSaml\Model\Protocol\StatusResponse;
 
@@ -24,18 +24,13 @@ use LightSaml\Model\Protocol\StatusResponse;
  */
 class LogoutController extends AbstractLogoutController
 {
-    /**
-     * @return SamlPluginInterface
-     */
-    protected function getSamlPlugin(): SamlPluginInterface
-    {
-        return Saml::getInstance();
-    }
+    use SamlPluginEnsured;
 
     /**
-     * @return ProviderInterface
+     * @param null $uid
+     * @return bool|ProviderInterface
      */
-    protected function getRemoteProvider($uid = null): ProviderInterface
+    protected function getRemoteProvider($uid = null)
     {
         $condition = [];
         if ($uid) {
@@ -51,10 +46,14 @@ class LogoutController extends AbstractLogoutController
      * @param ProviderInterface $provider
      * @throws \flipbox\saml\core\exceptions\InvalidMetadata
      * @throws \yii\base\ExitException
+     * @throws \yii\base\InvalidConfigException
      */
     protected function send(SamlMessage $samlMessage, ProviderInterface $provider)
     {
-        Factory::send($samlMessage, $provider);
+        Saml::getInstance()->getBindingFactory()->send(
+            $samlMessage,
+            $provider
+        );
         \Craft::$app->end();
     }
 
@@ -62,9 +61,12 @@ class LogoutController extends AbstractLogoutController
      * @param Request $request
      * @return StatusResponse
      * @throws \flipbox\saml\core\exceptions\InvalidSignature
+     * @throws \yii\base\InvalidConfigException
      */
     protected function receive(Request $request): StatusResponse
     {
-        return Factory::receive($request);
+        return Saml::getInstance()->getBindingFactory()->receive(
+            $request
+        );
     }
 }
