+++ src/controllers/MetadataController.php
@@ -9,16 +9,12 @@
 namespace flipbox\saml\sp\controllers;
 
 use flipbox\saml\core\controllers\AbstractMetadataController;
-use flipbox\saml\core\SamlPluginInterface;
 use flipbox\saml\sp\records\ProviderRecord;
-use flipbox\saml\sp\Saml;
+use flipbox\saml\sp\traits\SamlPluginEnsured;
 
 class MetadataController extends AbstractMetadataController
 {
-    protected function getSamlPlugin(): SamlPluginInterface
-    {
-        return Saml::getInstance();
-    }
+    use SamlPluginEnsured;
 
     /**
      * @inheritdoc
