+++ src/controllers/cp/view/metadata/EditController.php
@@ -9,22 +9,17 @@
 namespace flipbox\saml\sp\controllers\cp\view\metadata;
 
 use flipbox\saml\core\controllers\cp\view\metadata\AbstractEditController;
-use flipbox\saml\core\SamlPluginInterface;
 use flipbox\saml\sp\records\ProviderRecord;
-use flipbox\saml\sp\Saml;
+use flipbox\saml\sp\traits\SamlPluginEnsured;
 
 class EditController extends AbstractEditController
 {
 
+    use SamlPluginEnsured;
+
     /**
-     * @inheritdoc
+     * @return string
      */
-    protected function getSamlPlugin(): SamlPluginInterface
-    {
-        return Saml::getInstance();
-    }
-
-
     protected function getProviderRecord()
     {
         return ProviderRecord::class;
