+++ src/fields/ExternalIdentity.php
@@ -9,20 +9,13 @@ namespace flipbox\saml\sp\fields;
 use craft\base\ElementInterface;
 use craft\helpers\UrlHelper;
 use flipbox\saml\core\fields\AbstractExternalIdentity;
-use flipbox\saml\core\SamlPluginInterface;
-use flipbox\saml\sp\Saml;
+use flipbox\saml\sp\traits\SamlPluginEnsured;
 use yii\db\Query;
 
 class ExternalIdentity extends AbstractExternalIdentity
 {
 
-    /**
-     * @inheritdoc
-     */
-    protected function getSamlPlugin(): SamlPluginInterface
-    {
-        return Saml::getInstance();
-    }
+    use SamlPluginEnsured;
 
     /**
      * @param $value
