+++ src/helpers/ProviderHelper.php
@@ -1,0 +1,14 @@
+<?php
+
+/**
+ * @copyright  Copyright (c) Flipbox Digital Limited
+ */
+
+namespace flipbox\saml\sp\helpers;
+
+use flipbox\saml\core\helpers\ProviderHelper as BaseProviderHelper;
+
+class ProviderHelper extends BaseProviderHelper
+{
+
+}
