+++ src/services/bindings/Factory.php
@@ -8,14 +8,10 @@
 
 namespace flipbox\saml\sp\services\bindings;
 
-use flipbox\saml\core\SamlPluginInterface;
 use flipbox\saml\core\services\bindings\AbstractFactory;
-use flipbox\saml\sp\Saml;
+use flipbox\saml\sp\traits\SamlPluginEnsured;
 
 class Factory extends AbstractFactory
 {
-    public static function getSamlPlugin(): SamlPluginInterface
-    {
-        return Saml::getInstance();
-    }
+    use SamlPluginEnsured;
 }
