+++ src/services/bindings/HttpRedirect.php
@@ -10,9 +10,9 @@ namespace flipbox\saml\sp\services\bindings;
 
 use flipbox\saml\core\exceptions\InvalidIssuer;
 use flipbox\saml\core\records\ProviderInterface;
+use flipbox\saml\core\services\bindings\AbstractHttpRedirect;
 use flipbox\saml\sp\Saml;
 use LightSaml\Model\Assertion\Issuer;
-use flipbox\saml\core\services\bindings\AbstractHttpRedirect;
 
 class HttpRedirect extends AbstractHttpRedirect
 {
