+++ src/services/messages/AuthnRequest.php
@@ -11,20 +11,16 @@ namespace flipbox\saml\sp\services\messages;
 use craft\base\Component;
 use flipbox\keychain\records\KeyChainRecord;
 use flipbox\saml\core\helpers\SecurityHelper;
-use flipbox\saml\core\records\AbstractProvider;
 use flipbox\saml\core\records\ProviderInterface;
 use flipbox\saml\core\services\messages\SamlRequestInterface;
+use flipbox\saml\core\services\traits\Security;
 use flipbox\saml\sp\models\Settings;
 use flipbox\saml\sp\records\ProviderRecord;
 use flipbox\saml\sp\Saml;
-use flipbox\saml\core\services\traits\Security;
-use LightSaml\Credential\X509Certificate;
 use LightSaml\Helper;
 use LightSaml\Model\Assertion\Issuer;
 use LightSaml\Model\Protocol\AbstractRequest;
-use LightSaml\Model\Protocol\SamlMessage;
 use LightSaml\SamlConstants;
-use RobRichards\XMLSecLibs\XMLSecurityKey;
 use yii\base\Event;
 
 class AuthnRequest extends Component implements SamlRequestInterface
