+++ src/services/messages/LogoutRequest.php
@@ -2,18 +2,11 @@
 
 namespace flipbox\saml\sp\services\messages;
 
-use flipbox\saml\core\SamlPluginInterface;
 use flipbox\saml\core\services\messages\AbstractLogoutRequest;
-use flipbox\saml\sp\Saml;
+use flipbox\saml\sp\traits\SamlPluginEnsured;
 
 class LogoutRequest extends AbstractLogoutRequest
 {
 
-    /**
-     * @inheritdoc
-     */
-    public function getSamlPlugin(): SamlPluginInterface
-    {
-        return Saml::getInstance();
-    }
+    use SamlPluginEnsured;
 }
