+++ src/services/messages/LogoutResponse.php
@@ -8,21 +8,11 @@
 
 namespace flipbox\saml\sp\services\messages;
 
-use craft\base\Component;
-use flipbox\saml\core\SamlPluginInterface;
 use flipbox\saml\core\services\messages\AbstractLogoutResponse;
 use flipbox\saml\core\services\messages\SamlResponseInterface;
-use flipbox\saml\sp\Saml;
-use LightSaml\Model\Protocol\AbstractRequest;
-use LightSaml\Model\Protocol\LogoutRequest as LogoutResponseModel;
+use flipbox\saml\sp\traits\SamlPluginEnsured;
 
 class LogoutResponse extends AbstractLogoutResponse implements SamlResponseInterface
 {
-    /**
-     * @return SamlPluginInterface
-     */
-    protected function getSamlPlugin(): SamlPluginInterface
-    {
-        return Saml::getInstance();
-    }
+    use SamlPluginEnsured;
 }
