+++ src/services/messages/Metadata.php
@@ -8,47 +8,13 @@
 
 namespace flipbox\saml\sp\services\messages;
 
-use craft\base\Component;
-use flipbox\saml\core\helpers\UrlHelper;
-use flipbox\keychain\keypair\OpenSSL;
-use flipbox\keychain\records\KeyChainRecord;
-use flipbox\saml\core\exceptions\InvalidMetadata;
-use flipbox\saml\core\helpers\SerializeHelper;
-use flipbox\saml\core\records\AbstractProvider;
-use flipbox\saml\core\records\ProviderInterface;
-use flipbox\saml\core\SamlPluginInterface;
 use flipbox\saml\core\services\messages\AbstractMetadata;
 use flipbox\saml\core\services\messages\MetadataServiceInterface;
-use flipbox\saml\sp\models\Provider;
-use flipbox\saml\sp\records\ProviderRecord;
-use flipbox\saml\sp\transformers\MetadataTransformer;
-use LightSaml\Model\Metadata\EntityDescriptor;
-use LightSaml\Model\Metadata\SingleLogoutService;
-use LightSaml\Model\Metadata\SpSsoDescriptor;
-use LightSaml\Model\Metadata\AssertionConsumerService;
-use LightSaml\SamlConstants;
-use flipbox\saml\sp\Saml;
+use flipbox\saml\sp\traits\SamlPluginEnsured;
 
 class Metadata extends AbstractMetadata implements MetadataServiceInterface
 {
-
-    /**
-     * Deprecated Constants
-     * @see
-     */
-
-    /**
-     * @deprecated
-     */
-    const LOGIN_LOCATION = 'sso/login';
-    /**
-     * @deprecated
-     */
-    const LOGOUT_RESPONSE_LOCATION = 'sso/logout';
-    /**
-     * @deprecated
-     */
-    const LOGOUT_REQUEST_LOCATION = 'sso/logout/request';
+    use SamlPluginEnsured;
 
     /**
      * @return array
@@ -56,17 +22,5 @@ class Metadata extends AbstractMetadata implements MetadataServiceInterface
     public function getSupportedBindings()
     {
         return $this->supportedBindings;
-    }
-
-    /**
-     * Utils
-     */
-
-    /**
-     * @inheritdoc
-     */
-    protected function getSamlPlugin(): SamlPluginInterface
-    {
-        return Saml::getInstance();
     }
 }
