+++ tests/unit/SamlTest.php
@@ -5,6 +5,8 @@ namespace flipbox\saml\sp\tests;
 use Codeception\Test\Unit;
 use flipbox\saml\sp\Saml;
 use flipbox\saml\sp\services\Login;
+use flipbox\saml\sp\services\login\User;
+use flipbox\saml\sp\services\login\UserGroups;
 use flipbox\saml\sp\services\Provider;
 use flipbox\saml\sp\services\ProviderIdentity;
 use flipbox\saml\sp\services\bindings\HttpPost;
@@ -45,5 +47,7 @@ class SamlTest extends Unit
         $this->assertInstanceOf(Metadata::class, $this->module->getMetadata());
         $this->assertInstanceOf(Response::class, $this->module->getResponse());
         $this->assertInstanceOf(Session::class, $this->module->getSession());
+        $this->assertInstanceOf(User::class, $this->module->getUser());
+        $this->assertInstanceOf(UserGroups::class, $this->module->getUserGroups());
     }
 }
