--- tests/unit/services/CpTest.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace flipbox\saml\sp\tests\services;
-
-use Codeception\Test\Unit;
-
-class CpTest extends Unit
-{
-
-}
