+++ tests/unit/services/ProviderIdentityTest.php
@@ -4,12 +4,13 @@ namespace flipbox\saml\sp\tests\services;
 
 use Codeception\Test\Unit;
 use flipbox\saml\sp\records\ProviderIdentityRecord;
+use flipbox\saml\sp\Saml;
 use flipbox\saml\sp\services\ProviderIdentity;
 
 class ProviderIdentityTest extends Unit
 {
     public function testGetRecordClass()
     {
-        $this->assertEquals(ProviderIdentityRecord::class, (new ProviderIdentity)->getRecordClass());
+        $this->assertEquals(ProviderIdentityRecord::class, Saml::getInstance()->getProviderIdentityRecordClass());
     }
 }
