+++ tests/unit/services/ProviderTest.php
@@ -4,12 +4,13 @@ namespace flipbox\saml\sp\tests\services;
 
 use Codeception\Test\Unit;
 use flipbox\saml\sp\records\ProviderRecord;
+use flipbox\saml\sp\Saml;
 use flipbox\saml\sp\services\Provider;
 
 class ProviderTest extends Unit
 {
     public function testGetRecordClass()
     {
-        $this->assertEquals(ProviderRecord::class, (new Provider)->getRecordClass());
+        $this->assertEquals(ProviderRecord::class, Saml::getInstance()->getProviderRecordClass());
     }
 }
